+++ templates/base.html
@@ -20,8 +20,8 @@
     <header class="bg-white shadow-sm sticky top-0 z-10">
         <div class="container mx-auto p-4 flex justify-between items-center">
             <nav class="flex items-center gap-6">
-                <a href="{{ url_for('index') }}" class="text-lg font-bold text-gray-900 hover:text-blue-600">Home</a>
-                <a href="{{ url_for('list_people') }}" class="text-lg font-bold text-gray-900 hover:text-blue-600">Review People</a>
+                <a href="{{ url_for('tagging_bp.index') }}" class="text-lg font-bold text-gray-900 hover:text-blue-600">Home</a>
+                <a href="{{ url_for('people_bp.list_people') }}" class="text-lg font-bold text-gray-900 hover:text-blue-600">Review People</a>
             </nav>
             <div class="text-center">
                 <span class="font-semibold">{{ stats.named_people_count }}</span> People Identified |
