+++ templates/people_list.html
@@ -5,8 +5,8 @@
     <h2 class="font-semibold text-2xl mb-6 text-center">Identified People ({{ people|length }})</h2>
     <div class="grid grid-cols-2 md:grid-cols-3 lg:grid-cols-4 gap-4">
         {% for person in people %}
-        <a href="{{ url_for('person_details', person_name=person.person_name) }}" class="block text-center p-2 bg-gray-50 rounded-lg hover:bg-blue-100 hover:shadow-lg transition-all">
-            <img src="{{ url_for('get_face_thumbnail', face_id=person.face_id) }}" alt="Face of {{ person.person_name }}" class="w-24 h-24 object-cover rounded-full mx-auto mb-2 shadow-md">
+        <a href="{{ url_for('people_bp.person_details', person_name=person.person_name) }}" class="block text-center p-2 bg-gray-50 rounded-lg hover:bg-blue-100 hover:shadow-lg transition-all">
+            <img src="{{ url_for('tagging_bp.get_face_thumbnail', face_id=person.face_id) }}" alt="Face of {{ person.person_name }}" class="w-24 h-24 object-cover rounded-full mx-auto mb-2 shadow-md">
             <p class="font-semibold text-gray-800 truncate">{{ person.person_name }}</p>
             <p class="text-sm text-gray-500">{{ person.face_count }} faces</p>
         </a>
